Simple compile dockerfile
Running `docker build -t bootloader-build .` should result in the docker image building and a complete set of images generated. You can access the machine with a `docker run -it bootloader-build /bin/bash`. Files are not currently exposed outside the docker image... making this only useful for testing the repo atm.